+++ src/main/java/models/TeacherPayment.java
@@ -2,7 +2,6 @@ package models;
 
 
 import javax.persistence.*;
-import java.time.Year;
 
 @Entity
 @Table(name = "teacher_payments")
@@ -10,7 +9,7 @@ public class TeacherPayment {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column
-    private int id;
+    private Long id;
 
     @Column
     private float amount;
@@ -35,11 +34,11 @@ public class TeacherPayment {
     public TeacherPayment() {
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
