+++ src/main/java/models/Town.java
@@ -8,7 +8,7 @@ public class Town {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column
-    private int id;
+    private Long id;
 
     @Column
     private String title;
@@ -17,15 +17,14 @@ public class Town {
     @JoinColumn(name = "perfecture_id")
     private Perfecture perfectObj;
 
-
     public Town() {
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
